Keep independent proof checks within focused boundaries
Settlement claim, finality, and balance validation now have a dedicated module, while public actor error classification is split into small predicates. The refactor restores the repository's file and function limits without changing any proof decision.\n\nConstraint: Rust files must stay at or below 200 lines and functions at or below 25 lines\nRejected: Keep one near-limit verifier module | future tamper checks would immediately violate the size contract\nConfidence: high\nScope-risk: narrow\nReversibility: clean\nDirective: Add new proof domains as focused validators rather than expanding the bundle coordinator\nTested: cargo test -p kamn-e2e-harness --test independent_agent_transaction_verifier_contract --test independent_agent_transaction_actor_verifier_contract; cargo clippy -p kamn-e2e-harness --all-targets -- -D warnings\nNot-tested: full workspace gates follow in integration\nRelated: #7103
